% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{accessors}
\alias{accessors}
\alias{get_fitted_model}
\alias{get_fitted_model.default}
\alias{get_fitted_model.trending_fit}
\alias{get_formula}
\alias{get_formula.default}
\alias{get_formula.trending_model}
\alias{get_formula.trending_fit}
\alias{get_response}
\alias{get_response.default}
\alias{get_response.trending_model}
\alias{get_response.trending_fit}
\title{Accessors generics}
\usage{
get_fitted_model(x, ...)

\method{get_fitted_model}{default}(x, ...)

\method{get_fitted_model}{trending_fit}(x, ...)

get_formula(x, ...)

\method{get_formula}{default}(x, ...)

\method{get_formula}{trending_model}(x, ...)

\method{get_formula}{trending_fit}(x, ...)

get_response(x, ...)

\method{get_response}{default}(x, ...)

\method{get_response}{trending_model}(x, ...)

\method{get_response}{trending_fit}(x, ...)
}
\arguments{
\item{x}{An \R object.}

\item{...}{Not currently used by any methods.}
}
\value{
\itemize{
\item \code{get_formula()}: the formula used to model temporal trends.
\item \code{get_response()}: the response variable of the underlying model.
\item \code{get_fitted_model()}: The underlying fitted model.
}
}
\description{
These gen can be used to access information stored in
\code{\link{trending_model}}, \code{\link[=fit.trending_model]{trending_fit}} and
\code{\link[=predict.trending_fit]{trending_predict}} objects.
}
\examples{
x = rnorm(100, mean = 0)
y = rpois(n = 100, lambda = exp(1.5 + 0.5*x))
dat <- data.frame(x = x, y = y)

poisson_model <- glm_model(y ~ x , family = "poisson")
fitted_poisson <- fit(poisson_model, dat)
get_response(fitted_poisson)
get_formula(fitted_poisson)
get_fitted_model(fitted_poisson)

}
