% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{fit}
\alias{fit}
\alias{fit.trending_model}
\alias{trending_model_fit}
\alias{fit.list}
\alias{trending_model_fit_list}
\title{Fitting for trending_model objects}
\usage{
fit(x, data, ...)

\method{fit}{trending_model}(x, data, ...)

\method{fit}{list}(x, data, ...)
}
\arguments{
\item{x}{the output of functions \code{lm_model}, \code{glm_model}, \code{glm_nb_model}, or
brms_model}

\item{data}{A \code{data.frame} to be used to train the model.}

\item{...}{Additional arguments passed to underlying models.}
}
\description{
\code{\link[=fit]{fit()}} fits a model using the given data to obtain an object of type
\code{trending_model_fit} or  \code{trending_model_fit_list}.
}
