% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.r
\name{fit.trending_model}
\alias{fit.trending_model}
\title{Fit a trending model}
\usage{
\method{fit}{trending_model}(object, data, ...)
}
\arguments{
\item{object}{A \code{\link{trending_model}} object (i.e. the output from function
\code{lm_model}, \code{glm_model}, \code{glm.nb_model}, or \code{brm_model}).}

\item{data}{A data frame containing the data to fit.}

\item{...}{Not currently used.}
}
\value{
A \code{trending_fit} object which is a subclass of
\link[tibble:tibble]{tibble} with columns:
\itemize{
\item fitted_model: the resulting fit from calling the underlying model
directly, i.e.
\itemize{
\item \code{lm_model}: a fitted model object of class \code{\link[stats:lm]{lm}}
\item \code{glm_model}: a fitted model object of class \code{\link[stats:glm]{glm}}
\item \code{glm.nb_model}: a fitted model object of class \code{\link[MASS:glm.nb]{negbin}}
\item \code{brm_model}: An object of class \code{\link[brms:brm]{brmsfit}}
}

\code{NULL} if fitting fails.
\item fitting_warnings: any warnings generated during fitting
\item fitting_errors: any errors generated during fitting
}
}
\description{
\code{fit} will return the output from fitting the model to the specified
data.
}
\examples{
x = rnorm(100, mean = 0)
y = rpois(n = 100, lambda = exp(1.5 + 0.5*x))
dat <- data.frame(x = x, y = y)

poisson_model <- glm_model(y ~ x , family = poisson)
negbin_model <- glm.nb_model(y ~ x)

fit(poisson_model, dat)
fit(negbin_model, dat)

}
\author{
Tim Taylor
}
