% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_list.R
\name{fit_list}
\alias{fit_list}
\title{Fit a list of trending models}
\usage{
fit_list(models, data, as_tibble = FALSE)
}
\arguments{
\item{models}{A list of \code{\link{trending_model}} object (i.e. a list of the output
from functions \code{lm_model}, \code{glm_model}, \code{glm_nb_model}, or \code{brm_model}).}

\item{data}{A data frame containing the data to fit.}

\item{as_tibble}{Should the output be converted to a tibble subclass.}
}
\value{
If \code{as_tibble = TRUE} a \code{trending_fit_tbl} object which is a
\code{\link[tibble:tibble]{tibble}} subclass with with one row for each fitted
model and columns:
\itemize{
\item model_name (optional - only if input \code{models} is a named list)
\item result: the resulting fit from calling the underlying model
directly, i.e.
\itemize{
\item \code{lm_model}: a fitted model object of class \code{\link[stats:lm]{lm}}
\item \code{glm_model}: a fitted model object of class \code{\link[stats:glm]{glm}}
\item \code{glm_nb_model}: a fitted model object of class \code{\link[MASS:glm.nb]{negbin}}
\item \code{brm_model}: An object of class \code{\link[brms:brm]{brmsfit}}
}

\code{NULL} if fitting fails.
\item warnings: any warnings generated during fitting
\item errors: any errors generated during fitting
}

If \code{as_tibble = FALSE}, a \code{trending_fit_list} where each entry is a
\code{trending_fit} object.
}
\description{
\code{fit_list()} will return the output from fitting a list of models to the
specified data.
}
\examples{
x = rnorm(100, mean = 0)
y = rpois(n = 100, lambda = exp(1.5 + 0.5*x))
dat <- data.frame(x = x, y = y)

poisson_model <- glm_model(y ~ x , family = poisson)
negbin_model <- glm_nb_model(y ~ x)

fit_list(list(poisson_model, negbin_model), dat)
fit_list(list(pm = poisson_model, nm = negbin_model), dat)

}
\author{
Tim Taylor
}
