% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit.trending_model}
\alias{fit.trending_model}
\alias{trending_model_fit}
\title{Fitting and prediction}
\usage{
\method{fit}{trending_model}(x, data, ...)
}
\arguments{
\item{x}{the output of functions \code{lm_model}, \code{glm_model}, \code{glm_nb_model} or
\code{brms_model}.}

\item{data}{a \code{data.frame} to be used to fit the model.}

\item{...}{further arguments passed to other methods: \code{lm} for \code{lm_model},
\code{glm} for \code{glm_model}, \code{MASS::glm_nb} for \code{glm_nb_model}, \code{brms::brm} for
\code{brms_model}.

#' @return
\itemize{
\item \code{fit}: The fitted model.
}}
}
\description{
Fit a model using data to obtain a
\url{trending_model_fit} object.
}
