% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ok.R
\name{ok}
\alias{ok}
\alias{ok.default}
\alias{ok.trending_fit}
\alias{ok.trending_prediction}
\alias{is_ok}
\alias{is_ok.default}
\alias{is_ok.trending_fit}
\alias{is_ok.trending_prediction}
\title{Error handling generics}
\usage{
ok(x, warnings, ...)

\method{ok}{default}(x, warnings, ...)

\method{ok}{trending_fit}(x, warnings = TRUE, unnest = TRUE, ...)

\method{ok}{trending_prediction}(x, warnings = TRUE, unnest = TRUE, ...)

is_ok(x, warnings, ...)

\method{is_ok}{default}(x, warnings, ...)

\method{is_ok}{trending_fit}(x, warnings = TRUE, ...)

\method{is_ok}{trending_prediction}(x, warnings = TRUE, ...)
}
\arguments{
\item{x}{\R object}

\item{warnings}{Include results in output that triggered warnings but
not errors.  Defaults to \code{FALSE}.}

\item{...}{Not currently used.}

\item{unnest}{Should the warning and error columns be dropped and just the
outcome returned. Defaults to \code{TRUE}}
}
\value{
\itemize{
\item \code{ok()}: returns rows from an object that did not error (or optionally
produced a warning).
\item \code{is_ok}: returns TRUE or FALSE depending on whether the call succeeded
or errored (or optionally produced a warning)
}
}
\description{
\code{is_ok()} and \code{ok()} are generics used to filter successful outcomes from
those that errored or gave warnings.
}
\author{
Tim Taylor
}
