% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.fitted_and_predicted}
\alias{plot.fitted_and_predicted}
\title{Plot}
\usage{
\method{plot}{fitted_and_predicted}(
  x,
  date_axis,
  known_y,
  predicted_y = "pred",
  lower_y = "lower",
  upper_y = "upper",
  ...
)
}
\arguments{
\item{x}{A \code{\link[=fit_and_predict]{fit_and_predict()}} object.}

\item{date_axis}{Variable in x representing the date axis.}

\item{known_y}{Variable in x representing the known counts.}

\item{predicted_y}{Variable in x representing the predicted counts.}

\item{lower_y}{Variable in x representing the lower prediction interval.}

\item{upper_y}{Variable in x representing the upper prediction interval.}

\item{...}{Not currently used.}
}
\description{
Plot an object of class \code{fitted_and_predicted}.
}
