% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.trending_model_prediction}
\alias{plot.trending_model_prediction}
\title{Plot trending prediction intervals}
\usage{
\method{plot}{trending_model_prediction}(
  x,
  date_axis,
  y = "pred",
  lower = "lower-pi",
  upper = "upper-pi",
  fitted_data = NULL,
  fitted_y = NULL,
  ...
)
}
\arguments{
\item{x}{A \code{trending_model_prediction} object.}

\item{date_axis}{Variable in x representing the date axis.}

\item{y}{Variable in x representing the counts (predicted or estimated)}

\item{lower}{Variable in x representing the lower (pi/ci) interval.}

\item{upper}{Variable in x representing the upper (pi/ci) interval.}

\item{fitted_data}{The data used to fit the plot (optional).}

\item{fitted_y}{The count variable of the fitting data.}

\item{...}{Not currently used.}
}
\description{
Temporary plotting function for experimenting.
}
