% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotplot}
\alias{plotplot}
\title{Plot}
\usage{
plotplot(
  x,
  date_axis,
  known_y,
  predicted_y = "pred",
  lower_y = "lower",
  upper_y = "upper",
  facets = NULL,
  ...
)
}
\arguments{
\item{x}{data.frame.}

\item{date_axis}{Variable in x representing the date axis.}

\item{known_y}{Variable in x representing the known counts.}

\item{predicted_y}{Variable in x representing the predicted counts.}

\item{lower_y}{Variable in x representing the lower prediction interval.}

\item{upper_y}{Variable in x representing the upper prediction interval.}

\item{facets}{Variables to facet by.}

\item{...}{Not currently used.}
}
\description{
Temporary plotting function for experimenting.
}
