% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train.trending_model}
\alias{train.trending_model}
\title{Train a model}
\usage{
\method{train}{trending_model}(x, data, ...)
}
\arguments{
\item{x}{the output of functions \code{lm_model}, \code{glm_model}, \code{glm_nb_model} or
\code{brms_model}.}

\item{data}{a \code{data.frame} to be used to train the model.}

\item{...}{Not currently used.}
}
\description{
Train a model using data to obtain a
\url{trending_model_fit} object.
}
