% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trending_model_fit-accessors.R
\name{get_model.trendbreaker_model_fit}
\alias{get_model.trendbreaker_model_fit}
\alias{trendbreaker_model_fit-accessors}
\alias{trendbreaker_model_fit-class}
\title{Accessors for trendbreaker_model_fit objects}
\usage{
\method{get_model}{trendbreaker_model_fit}(x, ...)
}
\arguments{
\item{x}{the output of functions \code{lm_model}, \code{glm_model}, or
\code{glm_nb_model}}

\item{...}{further arguments passed to other methods}
}
\description{
These functions can be used to access information stored in
\code{trendbreaker_model_fit} objects. See details.
}
\details{
The following accessors are available:
\itemize{
\item \code{get_model()}: get the fitted model stored in the object
\item \code{predict()}: get model predictions for user-provided data, including
average predictions and prediction intervals
}
}
\author{
Thibaut Jombart, Dirk Schumacher
}
