% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_model}
\alias{get_model}
\alias{trendfit-generics}
\alias{get_formula}
\alias{get_response}
\alias{get_family}
\alias{fit}
\alias{fit_and_predict}
\title{S3 generics for trendfit}
\usage{
get_model(x, ...)

get_formula(x, ...)

get_response(x, ...)

get_family(x, ...)

fit(x, data, ...)

fit_and_predict(x, data, new_data, ...)
}
\arguments{
\item{x}{the object to access information from}

\item{...}{further arguments used in methods}

\item{data}{a \code{data.frame} to be used for fitting the model.}

\item{new_data}{a \code{data.frame} containing data for which predictions are to
be derived.}
}
\description{
These are generic functions used by the \emph{trendfit} package, mostly used for
accessing content of various objects. See \code{?trendfit-accessors} for methods
relating to \code{trendfit} objects, and \code{trendfit_model-accessors} for methods
relating to \code{trendfit_model} objects.
}
\seealso{
\url{trendfit-accessors},
\url{trendfit_model-accessors}
}
