% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trending-package.R
\docType{package}
\name{trendfit-package}
\alias{trendfit}
\alias{trendfit-package}
\title{trendfit: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/reconhub/trendfit}
  \item Report bugs at \url{https://github.com/reconhub/trendfit/issues}
}

}
\author{
\strong{Maintainer}: Tim Taylor \email{tim.taylor@hiddenelephants.co.uk} (\href{https://orcid.org/0000-0002-8587-7113}{ORCID}) [contributor]

Authors:
\itemize{
  \item Dirk Schumacher \email{mail@dirk-schumacher.net}
  \item Thibaut Jombart \email{thibautjombart@gmail.com}
}

}
\keyword{internal}
