% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trending_model-accessors.R
\name{get_formula.trendfit_model}
\alias{get_formula.trendfit_model}
\alias{trendfit_model-accessors}
\alias{trendfit_model-class}
\alias{get_response.trendfit_model}
\alias{get_family.trendfit_model}
\title{Accessors for trendfit_model objects}
\usage{
\method{get_formula}{trendfit_model}(x, ...)

\method{get_response}{trendfit_model}(x, ...)

\method{get_family}{trendfit_model}(x, ...)
}
\arguments{
\item{x}{the output of functions \code{lm_model}, \code{glm_model}, \code{glm_nb_model}, or
brms_model}

\item{...}{further arguments passed to other methods}
}
\description{
These functions can be used to access information stored in \code{trendfit_model}
objects. See details.
}
\details{
The following accessors are available:
\itemize{
\item \code{get_formula()}: get the formula used to model temporal trends
\item \code{get_response()}: get the name of the response variable
\item \code{get_family()}: get the model family, indicating the type of distribution
assumed for the response variable
}
}
\author{
Thibaut Jombart, Dirk Schumacher
}
