% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-interface.R
\name{glm_model}
\alias{glm_model}
\alias{trendfit_model}
\alias{trendfit_models}
\alias{glm_nb_model}
\alias{lm_model}
\alias{brms_model}
\alias{format.trendfit_model}
\alias{print.trendfit_model}
\title{Modeling interface}
\usage{
glm_model(formula, family)

glm_nb_model(formula)

lm_model(formula)

brms_model(formula, family)

\method{format}{trendfit_model}(x, ...)

\method{print}{trendfit_model}(x, ...)
}
\arguments{
\item{formula}{the formula of the model, with the response variable on the
left of a tilde symbol, and predictors on the right hand-side; variable
names used in the formula will need to be matched by columns in the \code{data}
input to other functions}

\item{family}{the model family to be used for the response variable}

\item{x}{an \code{trendfit_model} object}

\item{...}{Unused at this time.}
}
\value{
A \code{trendfit_model} object (S3 class inheriting \code{list}), containing
items which can be accessed by various accessors - see
\code{?trendfit_model-accessors}
}
\description{
These functions wrappers around various modelling tools to ensure a
consistent input for \emph{trendfit} functions. See details for available model
interfaces.
}
\details{
The following interfaces are available:
\itemize{
\item \code{lm_model}: interface for linear models implemented in
\url{stats::lm}
\item \code{glm_model}: interface for generalised linear models (GLMs) implemented in
\verb{[stats::glm](stats::glm)}
\item \code{glm_nb_model}: interface for negative binomial generalied linear models
implemented in \url{MASS::glm_nb}
\item \code{brms_model}: interface for Bayesian regression models implemented in
\url{brms::brm}
}
}
\author{
Dirk Schumacher
}
