% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trending-package.R
\docType{package}
\name{trending-package}
\alias{trending}
\alias{trending-package}
\title{trending: Model Temporal Trends}
\description{
Provides a coherent interface to multiple modelling tools for
  fitting trends along with a standardised approach for generating confidence
  and prediction intervals.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/reconhub/trending}
  \item Report bugs at \url{https://github.com/reconhub/trending/issues}
}

}
\author{
\strong{Maintainer}: Tim Taylor \email{tim.taylor@hiddenelephants.co.uk} (\href{https://orcid.org/0000-0002-8587-7113}{ORCID}) [contributor]

Authors:
\itemize{
  \item Dirk Schumacher \email{mail@dirk-schumacher.net}
  \item Thibaut Jombart \email{thibautjombart@gmail.com}
}

}
\keyword{internal}
