% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trending_model-accessors.R
\name{get_formula.trending_model}
\alias{get_formula.trending_model}
\alias{trending_model-accessors}
\alias{trending_model-class}
\alias{get_response.trending_model}
\alias{get_family.trending_model}
\alias{fit.trending_model}
\title{Accessors for trending_model objects}
\usage{
\method{get_formula}{trending_model}(x, ...)

\method{get_response}{trending_model}(x, ...)

\method{get_family}{trending_model}(x, ...)

\method{fit}{trending_model}(x, data, ...)
}
\arguments{
\item{x}{the output of functions \code{lm_model}, \code{glm_model}, \code{glm_nb_model}, or
brms_model}

\item{...}{further arguments passed to other methods}

\item{data}{a \code{data.frame} to be used to train the model#'}
}
\description{
These functions can be used to access information stored in \code{trending_model}
objects. See details.
}
\details{
The following accessors are available:
\itemize{
\item \code{get_formula()}: get the formula used to model temporal trends
\item \code{get_response()}: get the name of the response variable
\item \code{get_family()}: get the model family, indicating the type of distribution
assumed for the response variable
\item \code{fit()}: fit a model using data to obtain a
\url{trending_model_fit} object
}
}
\author{
Thibaut Jombart, Dirk Schumacher
}
