% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trending_model_accessors.R
\name{trending_model_accessors}
\alias{trending_model_accessors}
\alias{get_formula}
\alias{get_formula.trending_model}
\alias{get_response}
\alias{get_response.trending_model}
\title{Accessors for trending_model objects}
\usage{
get_formula(x, ...)

\method{get_formula}{trending_model}(x, ...)

get_response(x, ...)

\method{get_response}{trending_model}(x, ...)
}
\arguments{
\item{x}{Object of class \code{\link{trending_model_fit}} (i.e. the output from fitting
a \code{\link{trending_model}}).}

\item{...}{Not currently used by any methods.}
}
\description{
These functions can be used to access information stored in
\code{\link{trending_model}} objects. See details.
}
\details{
The following accessors are available:
\itemize{
\item \code{get_formula()}: get the formula used to model temporal trends;
\item \code{get_response()}: get the name of the response variable.
}
}
