% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fitting.r
\name{fit.trending_model}
\alias{fit.trending_model}
\alias{fit.list}
\title{Fitting for trending_model objects}
\usage{
\method{fit}{trending_model}(x, data, ...)

\method{fit}{list}(x, data, ...)
}
\arguments{
\item{x}{the output of functions \code{lm_model}, \code{glm_model}, \code{glm_nb_model}, or
brms_model}

\item{data}{A \code{data.frame} to be used to train the model.}

\item{...}{Additional arguments passed to underlying models.}
}
\description{
\code{\link[=fit]{fit()}} fits a model using the given data to obtain an object of type
\url{trending_model_fit} or
\url{trending_model_fit_list}
}
