% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-and-predict.R
\name{fit.trendfit_model}
\alias{fit.trendfit_model}
\alias{predict.trendfit_model_fit}
\alias{fit_and_predict.trendfit_model}
\title{Fitting and prediction}
\usage{
\method{fit}{trendfit_model}(x, data, ...)

\method{predict}{trendfit_model_fit}(object, new_data, alpha = 0.05, ...)

\method{fit_and_predict}{trendfit_model}(x, data, new_data, alpha = 0.05, ...)
}
\arguments{
\item{x}{the output of functions \code{lm_model}, \code{glm_model}, \code{glm_nb_model} or
\code{brms_model}.}

\item{data}{a \code{data.frame} to be used to fit the model.}

\item{...}{Not used.}

\item{object}{an \code{trendfit_model_fit} object}

\item{new_data}{a \code{data.frame} containing data for which predictions are to
be derived.}

\item{alpha}{the alpha threshold to be used for prediction intervals,
defaulting to 0.05, i.e. 95\% prediction intervals are derived.}
}
\value{
\itemize{
\item \code{predict}: data.frame with prediction and prediction intervals.
}

\itemize{
\item \code{fit_and_predict}: data.frame with both the values used to fit the model,
the prediction and the prediction intervals.
}
}
\description{
Fit a model using data to obtain a
\url{trendfit_model_fit} object.
}
