% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.trending_model_prediction}
\alias{plot.trending_model_prediction}
\title{Plot trending prediction intervals}
\usage{
\method{plot}{trending_model_prediction}(
  x,
  date_axis,
  predicted_y = "pred",
  lower_y = "lower",
  upper_y = "upper",
  fitted_data = NULL,
  fitted_y = NULL,
  ...
)
}
\arguments{
\item{x}{A \code{trending_model_prediction} object.}

\item{date_axis}{Variable in x representing the date axis.}

\item{predicted_y}{Variable in x representing the predicted counts.}

\item{lower_y}{Variable in x representing the lower prediction interval.}

\item{upper_y}{Variable in x representing the upper prediction interval.}

\item{fitted_data}{The data used to fit the plot (optional).}

\item{fitted_y}{The count variable of the fitting data.}

\item{...}{Not currently used.}
}
\description{
Temporary plotting function for experimenting.
}
