% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.trending_model_prediction}
\alias{plot.trending_model_prediction}
\title{Plot trending prediction intervals}
\usage{
\method{plot}{trending_model_prediction}(
  x,
  date_axis,
  counts = "pred",
  plot_ci = TRUE,
  plot_pi = TRUE,
  lower_ci = "lower-ci",
  upper_ci = "upper-ci",
  lower_pi = "lower-pi",
  upper_pi = "upper-pi",
  fitted_data = NULL,
  fitted_y = NULL,
  ...
)
}
\arguments{
\item{x}{A \code{trending_model_prediction} object.}

\item{date_axis}{Variable in x representing the date axis.}

\item{counts}{Variable in x representing the counts (predicted or estimated).}

\item{plot_ci}{Should confidence intervals be plotted (\code{TRUE}/\code{FALSE})}

\item{plot_pi}{Should prediction intervals be plotted (\code{TRUE}/\code{FALSE})}

\item{lower_ci}{Variable in x representing the lower confidence interval.}

\item{upper_ci}{Variable in x representing the upper confidence interval.}

\item{lower_pi}{Variable in x representing the lower prediction interval.}

\item{upper_pi}{Variable in x representing the upper prediction interval.}

\item{fitted_data}{The data used to fit the plot (optional).}

\item{fitted_y}{The count variable of the fitting data.}

\item{...}{Not currently used.}
}
\description{
Temporary plotting function for experimenting.
}
