% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_model}
\alias{get_model}
\alias{trending-generics}
\alias{get_formula}
\alias{get_response}
\alias{fit}
\title{S3 generics for trending}
\usage{
get_model(x, ...)

get_formula(x, ...)

get_response(x, ...)

fit(x, data, ...)
}
\arguments{
\item{x}{the object to access information from}

\item{...}{further arguments used in methods}

\item{data}{a \code{data.frame} to be used for fitting the model.}
}
\description{
These are generic functions used by the \emph{trending} package, mostly used for
accessing content of various objects. See \code{?trending-accessors} for methods
relating to \code{trending} objects, and \code{trending_model-accessors} for methods
relating to \code{trending_model} objects.
}
\seealso{
\url{trending-accessors},
\url{trending_model-accessors}
}
