% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trending_model_fit-accessors.R
\name{get_model.trending_model_fit}
\alias{get_model.trending_model_fit}
\title{Accessors for trending_model_fit objects}
\usage{
\method{get_model}{trending_model_fit}(x, ...)
}
\arguments{
\item{x}{the output of functions \code{lm_model}, \code{glm_model}, \code{glm_nb_model} or
\code{brms_model}}

\item{...}{Not currently used}
}
\description{
These functions can be used to access information stored in
\code{trending_model_fit} objects. See details.
}
\details{
The following accessors are available:
\itemize{
\item \code{get_model()}: get the fitted model stored in the object
}
}
\author{
Thibaut Jombart, Dirk Schumacher
}
