% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-prediction.R
\name{predict.trending_model_fit}
\alias{predict.trending_model_fit}
\alias{predict.trending_model_fit_glm}
\title{Predict methods}
\usage{
\method{predict}{trending_model_fit}(
  object,
  new_data,
  alpha = 0.05,
  interval = c("both", "ci", "pi", "none"),
  ...
)

\method{predict}{trending_model_fit_glm}(
  object,
  new_data,
  alpha = 0.05,
  interval = c("both", "ci", "pi", "none"),
  uncertain = TRUE,
  ...
)
}
\arguments{
\item{object}{A \code{trending_model_fit} or \code{trending_model_fit_list} object.}

\item{new_data}{A \code{data.frame} containing data for which predictions are to
be derived.}

\item{alpha}{The alpha threshold to be used for prediction intervals,
defaulting to 0.05, i.e. 95\% prediction intervals are derived.}

\item{interval}{Which interval to add to the data.  Can be one of \code{ci}
(confidence interval), \code{pi} (prediction interval), \code{both} (both intervals)
or \code{none} (no intervals added).}

\item{uncertain}{Default TRUE.  If FALSE uncertainty in the fitted paramaters
is ignored when generating the prediction intervals.  These will lead to
narrower intervals.}
}
\description{
These functions can be used to generated estimated values and associated
confidence/prediction intervals for trending_model_fit objects.
}
