% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-prediction.R
\name{trending_model_fit-prediction}
\alias{trending_model_fit-prediction}
\alias{predict.trending_model_fit}
\alias{predict.trending_model_fit_list}
\title{Predict methods}
\usage{
\method{predict}{trending_model_fit}(object, new_data, alpha = 0.05, add_pi = TRUE, uncertain = TRUE, ...)

\method{predict}{trending_model_fit_list}(object, new_data, alpha = 0.05, add_pi = TRUE, uncertain = TRUE, ...)
}
\arguments{
\item{object}{A \code{trending_model_fit} or \code{trending_model_fit_list} object.}

\item{new_data}{A \code{data.frame} containing data for which predictions are to
be derived.}

\item{alpha}{The alpha threshold to be used for prediction intervals,
defaulting to 0.05, i.e. 95\% prediction intervals are derived.}

\item{add_pi}{Add a prediction interval to the output. Default TRUE.}

\item{uncertain}{Only used for glm models.  Default TRUE.  If FALSE
uncertainty in the fitted paramaters is ignored when generating the
prediction intervals.}

\item{...}{Not currently used.}
}
\description{
These functions can be used to generated estimated values and associated
confidence/prediction intervals for trending_model_fit objects.
}
